requirements.txt
alembic==1.10.4
aniso8601==9.0.1
anyio==3.6.2
apispec==6.3.0
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.2.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
cloud-init==22.2
colorama==0.4.6
comm==0.1.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
flask-restx==1.0.3
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
fonttools==4.39.4
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==16.10.1
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==4.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.24.3
oauthlib==3.1.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
Pygments==2.15.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.4
pytest==7.2.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==5.3.1
pyzmq==25.0.2
qrcode==7.3
requests==2.22.0
requests-unixsocket==0.2.0
reverse==0.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.10.1
seaborn==0.12.2
SecretStorage==2.3.1
Send2Trash==1.8.2
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
sos==4.3
soupsieve==2.4.1
SQLAlchemy==2.0.12
ssh-import-id==5.10
stack-data==0.6.2
statsmodels==0.14.0
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
Twisted==18.9.0
typing-extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==27.10
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.20.0
wadllib==1.3.3
wcwidth==0.2.6
webargs==8.2.0
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
zipp==3.15.0
zope.interface==4.7.1
